Add `Query` interface to DBs
As we go forward, we'll need more and more exotic filters to pull alerts
out of a database. We could do this by calling `GetAlerts` and then filtering
ourselves, but certain DBs could provide optimisations for different queries
such as the inmemory db being able to shortcut exact labelset matches.

This commit removes the `GetExistingAlert` method from DBs and introduces a
new method `QueryAlerts` along side a corresponding `AlertQuery` interface.
This provides a mechanism for consumers to generically filter alerts as necessary
and allows DBs to optimise queries as they need.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>