Add in tests for NotifyService
This adds in tests for the NotifyService. In doing so, we have to do a bit
of rearranging. In particular, we move internal/services/notify/notifier.go
into its own package to avoid an import loop with the mock package, and we
introduce a new `stubs` package that wraps time.Now() into a var that can
be overriden in tests.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>