Add initial support for ack validation
This adds in an initial support for validation pipelines. This allows
things like specifying arbitary filters to be applied to data before
it comes into the system.

In particular, this implements a filter infrastructure for Alert Acknowledgements,
which allows enforcing data rules against acks, like that they must have a valid email.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>