Add silence applying to the event delegate
This adds in logic to move alerts into the AlertStatusSilenced state when
they come in and theres a matching silence. This logic, and its opposite -
applying a new silence when it comes in, are applied in the eventdelegate.
It has to be inline like this to ensure that we apply silence logic _before_
the alert notifier has a chance to see it.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>