Refactor filter logic
Previously, we had seperate implementations of filters for each type -
alerts, acks etc all had to be implemented seperatly. This changes filters
up so that they all have one method `Filter` that takes a `Fielder`. Fielders
can provide arbitrary fields for the filter to filter on.

We then implement Fielder for both alerts and alert acknowledgements,
to make this a noop.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>