Update AlertAcknowledgementIsValid to return an error
In order to tell a user _why_ an alert acknowledgement is invalid, we
need to return an error. This commit updates the function to return an
error, and updates the tests to check for the error.

The logic here is that we traverse the ack tree, collecting errors at
the end of every branch. If we eventually reach the end of the tree
without an error, we return nil. Otherwise we return a multierror
of every node that we got rejected by.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>