Make data validation more generic
Data validation previously was only used for validating AlertAcknowledgements.
Instead of implementing seperate validation for each type of data, this
changes the Config method to be a single one that takes a config.Fielder.
This allows us to use the same validation pipeline for all types of data,
without having to implement a seperate method for each type we want to validate.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>