Add DurationFilter
Here we add a filter than can be used to filter out models that have durations.
The only model that provides a duration at the moment is the Silences,
so this allows validating properties of a silence. e.g. enforcing the use
of a JIRA ticket in the comment if a silence is long.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>