Add initial stats support
Stats queries will allow pulling numbers out of the alerts database,
such as the number of alerts in a given state, or the number of alerts
altogether, grouped by start time. This will allow us to build graphs
and dashboards in the UI. This commit adds the initial support for
stats queries, and adds a test to ensure that the query is parsed
correctly.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>