Generate and embed frontend assets in go build
This adds in a new package with go:generate directives that run the frontend build,
and then embeds the resulting assets into the go binary. This means that we can ship
a single binary that contains both the frontend and backend, and can be run without
needing to install nodejs or npm.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>