Buffer events in the DBEventDelegate
Becaues the DBEventDelegate receives alerts and silences from the clusterer
one at a time, the DBEventDelegate passes them to the DB one at a time.
This is slow because the DB can unlock all sorts of performance improvements
with batching, like using a single transaction for multiple inserts.

This commit adds a buffer to the DBEventDelegate, which batches events
together and passes them to the DB in batches. This improves performance
significantly on databases with high latency like the new boltdb.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>