Don't double lock in buffer db
Locking in both Flush and flushAlerts causes a deadlock. This is
because the lock is held in Flush, and then flushAlerts is called
which tries to acquire the lock again. This is not needed as the
lock is already held in Flush. This commit removes the lock from
flushAlerts and adds a comment to indicate that the lock must be
held before calling it.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>