Add in a few config datastructures
This adds in a few config datastructures that will be used to start implementing
more interesting notifiers. In particular, we introduce a `Secret` wrapper
that will be used to wrap sensitive data in the config file, and a `MaybeFile`
wrapper that will be used to optionally load a value from a literal or a
seperate file.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>