Update frontend to use options instead of arguments
As we add more options to the API, calling it with arguments becomes
more and more unwieldy. This commit updates the frontend to use options
instead of arguments, which makes it easier to add new options, and makes
existing calls easier to read by naming their arguments and removing
invalid ones.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>