Fix frontend embedding
This fixes how we embed the frontend into the binary. When rearranging
things earlier, we neglected to remove the build/ from the fs.Sub, and
the method that we were using in the frontend to detect a dev env vs
a prod env was broken. This fixes both of those, and adds in an integration
test to ensure that the frontend is being embedded correctly.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>