Finish off initial create silence flow
This finishes off the initial UI for creating a silence. We split the
silence creation into two steps, the first is the form to enter the
silence details, the second is a preview of the silence. This allows a
user to review the impact of a silence before its created.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>