Allow filters on the GET silences api
This adds the ability to filter the silences returned by the GET
silences API. This is done by adding a new query parameter, `filter`,
which is a set of matchers that are ANDed together. This is the same
format as the filter parameter on the GET alerts API, although the semantics
are slightly different. Whereas the alerts API will generally apply the matchers
to the _labels_ of alerts, silences are matched literally against the matchers
in each silence - no special treatment is given to negative or regex matchers.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>