Unify filter errors
Previously, filters would return a bool, which, if false, would trigger
a call to `Describe` on that filter to describe the error. This makes
nice error reporting hard because once the Describe call is made, most
of the context is lost.

This changes that interface so that the `Filter` call returns an error
directly, and the `Describe` call is removed. This means that the
Filter can return an error with all the context that it has.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>