Allow validation of alerts
Previously, we just accepted any alert that was sent to us. This
adds in a call to `ValidateData` on the alert, which will check
that all the alerts pass validation before being sent through to the DB.

Because we have no mechanism for signalling a partial failure, we
just return an error if any of the alerts fail validation.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>